:recycle: Refactoring
