+++ frontend/src/components/PreviewHeader/PreviewHeader.tsx
@@ -1,0 +1,3 @@
+export const PreviewHeader = (): JSX.Element => {
+  return <h2>Preview</h2>;
+};
