+++ frontend/src/interfaces/TrophySVGIconsProps.ts
@@ -1,0 +1,3 @@
+export type TrophySVGIconsProps = {
+  trophies: string;
+};
